requirements.txt
# requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
aiohttp==3.9.0
aiosqlite==0.19.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx
